Refactor status to use sum type
